requirements.txt
openai>=1.3.0
notion-client>=2.2.1
python-dateutil>=2.8.2
PyYAML>=6.0.1
requests>=2.31.0
pydantic>=2.5.0